# swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/account_details/:stash_account": {
      "get": {
        "description": "This will show latest account balances, latest identity associated with the account (display name, email, social medias, etc), and all balance related events.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Gets latest account details",
        "operationId": "getAccountDetails",
        "parameters": [
          {
            "type": "string",
            "description": "StashAccount",
            "name": "StashAccount",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/AccountDetailsView"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          }
        }
      }
    },
    "/account_rewards/:stash_account": {
      "get": {
        "description": "This will show rewards for account for given time period from \"start\" to \"end\", If \"end\" is not specified,\nwill return rewards until most recently indexed block",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Gets rewards for account for time period",
        "operationId": "getAccountRewards",
        "parameters": [
          {
            "type": "string",
            "description": "Address",
            "name": "Address",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "format": "date-time",
            "description": "Start",
            "name": "Start",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "format": "date-time",
            "description": "End",
            "name": "End",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/AccountRewardsView"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          }
        }
      }
    },
    "/block": {
      "get": {
        "description": "If no height is provided, will return last indexed block",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Gets block details and extrinsics for given height",
        "operationId": "getBlockByHeight",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "Height",
            "name": "Height",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/BlockDetailsView"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          }
        }
      }
    },
    "/block_times/:limit": {
      "get": {
        "description": "Gets stats related to how fast blocks are being produced in the chain",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "getBlockTimes",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "Limit",
            "name": "Limit",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/BlockTimesView"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          }
        }
      }
    },
    "/blocks_summary": {
      "get": {
        "description": "Gets summary for blocks",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Gets blocks summary",
        "operationId": "getBlocksSummary",
        "parameters": [
          {
            "type": "string",
            "description": "Interval",
            "name": "Interval",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "Period",
            "name": "Period",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/BlockSummary"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          }
        }
      }
    },
    "/rewards/:stash_account": {
      "get": {
        "description": "This will show all rewards (claimed and unclaimed) for account for given eras from \"start\" to \"end\" eras. If \"end\" is not specified,\nwill return rewards until most recently indexed block.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Gets rewards for account for eras",
        "operationId": "getRewards",
        "parameters": [
          {
            "type": "string",
            "description": "StashAccount",
            "name": "StashAccount",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "description": "Start",
            "name": "Start",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "description": "End",
            "name": "End",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/RewardsForErasView"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          }
        }
      }
    },
    "/status": {
      "get": {
        "description": "This will show latest status of the chain/node and indexer",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Gets latest status",
        "operationId": "getStatus",
        "parameters": [
          {
            "type": "boolean",
            "description": "IncludeChainStatus",
            "name": "IncludeChainStatus",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ChainDetailsView"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          }
        }
      }
    },
...
    "/transactions": {
      "get": {
        "description": "This will show all signed transactions for height. If height not specified, returns transactions\nfor most recently synced height.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "Gets all signed transactions for height",
        "operationId": "getTransactions",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "Height",
            "name": "Height",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TransactionsView"
          },
          "400": {
            "$ref": "#/responses/BadRequestResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Account": {
      "type": "object",
      "properties": {
        "fee_frozen": {
          "type": "string",
          "x-go-name": "FeeFrozen"
        },
        "free": {
          "type": "string",
          "x-go-name": "Free"
        },
        "misc_frozen": {
          "type": "string",
          "x-go-name": "MiscFrozen"
        },
        "reserved": {
          "type": "string",
          "x-go-name": "Reserved"
        }
      },
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/usecase/account"
    },
    "BalanceDeposit": {
      "type": "object",
      "properties": {
        "amount": {
          "description": "Amount is balance that was deposited into account",
          "type": "string",
          "x-go-name": "Amount"
        },
        "height": {
          "description": "Height is block height transaction occured",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Height"
        },
        "transaction_hash": {
          "description": "Hash is transaction hash",
          "type": "string",
          "x-go-name": "Hash"
        }
      },
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/usecase/account"
    },
    "BalanceTransfer": {
      "type": "object",
      "properties": {
        "amount": {
          "description": "Amount is balance that was transferred",
          "type": "string",
          "x-go-name": "Amount"
        },
        "height": {
          "description": "Height is block height transaction occured",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Height"
        },
        "kind": {
          "description": "Kind is transfer kind (either \"in\" for transfer into account or \"out\" for transfer out of account)",
          "type": "string",
          "x-go-name": "Kind",
          "example": "in"
        },
        "participant": {
          "description": "Participant is account",
          "type": "string",
          "x-go-name": "Participant"
        },
        "transaction_hash": {
          "description": "Hash is transaction hash",
          "type": "string",
          "x-go-name": "Hash"
        }
      },
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/usecase/account"
    },
    "Bonded": {
      "type": "object",
      "properties": {
        "amount": {
          "description": "Amount is balance that was bonded",
          "type": "string",
          "x-go-name": "Amount"
        },
        "height": {
          "description": "Height is block height transaction occured",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Height"
        },
        "receiver": {
          "description": "Receiver is account that bonded",
          "type": "string",
          "x-go-name": "Receiver"
        },
        "transaction_hash": {
          "description": "Hash is transaction hash",
          "type": "string",
          "x-go-name": "Hash"
        }
      },
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/usecase/account"
    },
    "Delegation": {
      "type": "object",
      "properties": {
        "amount": {
          "$ref": "#/definitions/Quantity"
        },
        "era": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Era"
        },
        "validator_stash": {
          "type": "string",
          "x-go-name": "ValidatorStash"
        }
      },
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/usecase/common"
    },
...
    "Quantity": {
      "description": "Quantity is a wrapper for big.Int",
      "type": "object",
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/types"
    },
...
    "Time": {
      "type": "object",
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/types"
    },
    "Unbonded": {
      "type": "object",
      "properties": {
        "amount": {
          "description": "Amount is balance that was unbonded",
          "type": "string",
          "x-go-name": "Amount"
        },
        "height": {
          "description": "Height is block height transaction occured",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Height"
        },
        "transaction_hash": {
          "description": "Hash is transaction hash",
          "type": "string",
          "x-go-name": "Hash"
        }
      },
      "x-go-package": "github.com/figment-networks/polkadothub-indexer/usecase/account"
    },
...
  },
  "responses": {
    "AccountDetailsView": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Account"
      },
      "headers": {
        "account": {},
        "address": {
          "type": "string",
          "description": "Address of account"
        },
        "bonded": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Bonded"
          },
          "description": "Bonded is a list of all staking.Bonded events for account"
        },
        "delegations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Delegation"
          },
          "description": "Delegations is a list of latest delegations for account"
        },
        "deposit": {
          "type": "string"
        },
        "deposits": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BalanceDeposit"
          },
          "description": "Deposits is a list of all balances.Deposit events for account"
        },
        "display_name": {
          "type": "string"
        },
        "email_name": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "legal_name": {
          "type": "string"
        },
        "riot_name": {
          "type": "string"
        },
        "transfers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BalanceTransfer"
          },
          "description": "Transfers is a list of all balances.Transfer events for account"
        },
        "twitter_name": {
          "type": "string"
        },
        "unbonded": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Unbonded"
          },
          "description": "Unbonded is a list of all staking.Unbonded events for account"
        },
        "web_name": {
          "type": "string"
        },
        "withdrawn": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Withdrawn"
          },
          "description": "Withdrawn is a list of all staking.Withdrawn events for account"
        }
      }
    },
    "AccountRewardsView": {
      "description": "",
      "headers": {
        "account": {
          "type": "string",
          "description": "Account is stash account"
        },
        "end_time": {
          "type": "string",
          "format": "date-time",
          "description": "End is end time of query period"
        },
        "rewards": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Reward"
          },
          "description": "Rewards is list of rewards"
        },
        "start_time": {
          "type": "string",
          "format": "date-time",
          "description": "Start is start time of query period"
        },
        "total_amount": {
          "type": "string",
          "description": "TotalAmount is summed total of rewards earned"
        }
      }
    },
    "BadRequestResponse": {
      "description": "",
      "schema": {
        "type": "object"
      },
      "headers": {
        "error": {},
        "status": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "BlockDetailsView": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Time"
      },
      "headers": {
        "extrinsics": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ExtrinsicDetailsView"
          }
        },
        "extrinsics_root": {
          "type": "string"
        },
        "hash": {
          "type": "string"
        },
        "height": {
          "type": "integer",
          "format": "int64"
        },
        "parent_hash": {
          "type": "string"
        },
        "state_root": {
          "type": "string"
        },
        "time": {}
      }
    },
    "BlockSummary": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Time"
      },
      "headers": {
        "block_time_avg": {
          "type": "number",
          "format": "double"
        },
        "count": {
          "type": "integer",
          "format": "int64"
        },
        "created_at": {},
        "id": {},
        "index_version": {
          "type": "integer",
          "format": "int64"
        },
        "time_bucket": {},
        "time_interval": {},
        "updated_at": {}
      }
    },
    "BlockTimesView": {
      "description": "",
      "headers": {
        "avg": {
          "type": "number",
          "format": "double"
        },
        "count": {
          "type": "integer",
          "format": "int64"
        },
        "diff": {
          "type": "number",
          "format": "double"
        },
        "end_height": {
          "type": "integer",
          "format": "int64"
        },
        "end_time": {
          "type": "string"
        },
        "start_height": {
          "type": "integer",
          "format": "int64"
        },
        "start_time": {
          "type": "string"
        }
      }
    },
    "ChainDetailsView": {
      "description": "",
      "schema": {
        "type": "object",
        "additionalProperties": {
          "type": "string"
        },
        "$ref": "#/definitions/Time"
      },
      "headers": {
        "app_name": {
          "type": "string"
        },
        "app_version": {
          "type": "string"
        },
        "chain_name": {
          "type": "string"
        },
        "chain_spec_version": {
          "type": "string"
        },
        "chain_type": {
          "type": "string"
        },
        "client_info": {
          "type": "string"
        },
        "genesis_hash": {
          "type": "string"
        },
        "go_version": {
          "type": "string"
        },
        "indexing_lag": {
          "type": "integer",
          "format": "int64",
          "description": "Lag is how many blocks the indexer is behind the current head of the chain"
        },
        "indexing_started": {
          "type": "boolean",
          "description": "IndexingStarted is false if indexer has never run before, otherwise returns true"
        },
        "last_index_version": {
          "type": "integer",
          "format": "int64",
          "description": "LastIndexVersion is config version of last indexed block"
        },
        "last_indexed_at": {
          "description": "LastIndexedAt is last time a block was indexed"
        },
        "last_indexed_era": {
          "type": "integer",
          "format": "int64",
          "description": "LastIndexedEra is era of last indexed block"
        },
        "last_indexed_era_height": {
          "type": "integer",
          "format": "int64",
          "description": "LastIndexedEraHeight is last indexed height where era was last in era"
        },
        "last_indexed_height": {
          "type": "integer",
          "format": "int64",
          "description": "LastIndexedHeight is height of last indexed block"
        },
        "last_indexed_session": {
          "type": "integer",
          "format": "int64",
          "description": "LastIndexedSession is session of last indexed block"
        },
        "last_indexed_session_height": {
          "type": "integer",
          "format": "int64",
          "description": "LastIndexedSessionHeight is last indexed height where session was last in session"
        },
        "last_indexed_time": {
          "description": "LastIndexedTime is last indexed block time"
        },
        "node_health": {
          "type": "string"
        },
        "node_local_peer_uid": {
          "type": "string"
        },
        "node_name": {
          "type": "string"
        },
        "node_properties": {},
        "node_roles": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "node_version": {
          "type": "string"
        }
      }
    },
...
    "RewardsForErasView": {
      "description": "",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/RewardEraSeq"
        }
      }
    },
...
    "TransactionsView": {
      "description": "",
      "headers": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ListItem"
          }
        }
      }
    },
...
  }
}